requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
dateparser>=1.1.0
pytz>=2023.3
